mesa: fix building with llvm
Use glslang's bin output in nativeBuildInputs to prevent propagating the "dev" output which
then injects the glslang libraries for host into target. Switch to using
the spirv-tools package for buildInputs instead of glslang to further
prevent the wrong things from propagating on host since we only need the
pkg-config and static libraries.